+++ windows/DMShot/Platform/IClipboardService.cs
@@ -2,6 +2,8 @@ using System.Drawing;
 namespace DMShot.Platform;
 public interface IClipboardService
 {
+    bool ContainsImage();
+    Bitmap? GetImage();
     void SetImage(Bitmap bmp);
     void SetGif(byte[] gifBytes, string gifFilePath);
 }